backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
Pillow==10.1.0

